requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
pytz==2023.3
requests==2.31.0
bleach==6.1.0
openpyxl==3.1.2
pdfkit==1.0.0
TForms==3.1.2